requirements.txt
# Основные библиотеки для Beast Lord Bot
pure-python-adb>=0.3.0.dev0      # Управление ADB без зависимостей
opencv-python>=4.8.0        # Распознавание изображений и шаблонов
Pillow>=10.0.0              # Работа с изображениями
schedule>=1.2.0             # Планировщик задач
loguru>=0.7.0               # Удобное логирование
PyYAML>=6.0                 # Работа с YAML конфигами
rich>=14.1.0

# Дополнительные утилиты
python-dateutil>=2.8.2     # Работа с датами для прайм-таймов
typing-extensions>=4.5.0   # Поддержка типизации